requirements.txt
discord.py
aiohttp
python-dotenv
flake8
